move primitives::constants::* into primitives module
There is no reason to have a submodule for constants. It is obvious that
they are constants.
