+++ src/util/merkleblock.rs
@@ -58,7 +58,7 @@ use std::io;
 use hashes::Hash;
 use hash_types::{Txid, TxMerkleNode};
 
-use primitives::constants::{MAX_BLOCK_WEIGHT, MIN_TRANSACTION_WEIGHT};
+use primitives::{MAX_BLOCK_WEIGHT, MIN_TRANSACTION_WEIGHT};
 use consensus::encode::{self, Decodable, Encodable};
 use util::merkleblock::MerkleBlockError::*;
 use {Block, BlockHeader, Transaction};
